+++ src/steps/account/converter.ts
@@ -9,17 +9,14 @@ export function createAccountEntity(): Entity {
   return createIntegrationEntity({
     entityData: {
       source: {
-        id: 'acme-unique-account-id',
-        name: 'Example Co. Acme Account',
+        id: 'kandji-account',
+        name: 'Kandji Account',
       },
       assign: {
-        _key: 'acme-unique-account-id',
+        _key: 'kandji-account',
         _type: Entities.ACCOUNT._type,
         _class: Entities.ACCOUNT._class,
-        mfaEnabled: true,
-        // This is a custom property that is not a part of the data model class
-        // hierarchy. See: https://github.com/JupiterOne/data-model/blob/master/src/schemas/Account.json
-        manager: 'Manager Name',
+        name: 'Kandji Account',
       },
     },
   });