+++ src/steps/account/index.ts
@@ -4,10 +4,8 @@ import {
 } from '@jupiterone/integration-sdk-core';
 
 import { IntegrationConfig } from '../../config';
-import { Steps, Entities } from '../constants';
+import { ACCOUNT_ENTITY_KEY, IntegrationSteps, Entities } from '../constants';
 import { createAccountEntity } from './converter';
-
-export const ACCOUNT_ENTITY_KEY = 'entity:account';
 
 export async function fetchAccountDetails({
   jobState,
@@ -19,7 +17,7 @@ export async function fetchAccountDetails({
 
 export const accountSteps: IntegrationStep<IntegrationConfig>[] = [
   {
-    id: Steps.ACCOUNT,
+    id: IntegrationSteps.ACCOUNT,
     name: 'Fetch Account Details',
     entities: [Entities.ACCOUNT],
     relationships: [],