+++ test/config.ts
@@ -10,12 +10,12 @@ if (process.env.LOAD_ENV) {
     path: path.join(__dirname, '../.env'),
   });
 }
-const DEFAULT_CLIENT_ID = 'dummy-acme-client-id';
-const DEFAULT_CLIENT_SECRET = 'dummy-acme-client-secret';
+const DEFAULT_API_URL = 'https://dummyurl.clients.us-1.kandji.io/api/v1/';
+const DEFAULT_ACCESS_TOKEN = 'dummy-kandji-access-token';
 
 export const integrationConfig: IntegrationConfig = {
-  clientId: process.env.CLIENT_ID || DEFAULT_CLIENT_ID,
-  clientSecret: process.env.CLIENT_SECRET || DEFAULT_CLIENT_SECRET,
+  apiUrl: process.env.API_URL || DEFAULT_API_URL,
+  accessToken: process.env.ACCESS_TOKEN || DEFAULT_ACCESS_TOKEN,
 };
 
 export function buildStepTestConfigForStep(stepId: string): StepTestConfig {