+++ CHANGELOG.md
@@ -7,3 +7,13 @@ and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+- Ingest new entities
+  - `kandji_account`
+  - `kandji_app`
+  - `kandji_device`
+  - `kandji_user`
+- Build new relationships
+  - `kandji_account_has_device`
+  - `kandji_device_has_app`
+  - `kandji_device_has_user`